% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker.R
\name{marker}
\alias{marker}
\title{Mark a submission}
\usage{
marker(submission, valid)
}
\arguments{
\item{submission}{valid submission}

\item{valid}{list of valid answers}
}
\value{
Prints True or False to the console
}
\description{
Mark a submission
}
\examples{
marker('a', letters)

}
