% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owhaley-package.R
\docType{package}
\name{owhaley-package}
\alias{owhaley}
\alias{owhaley-package}
\title{owhaley: Whaley good messages}
\description{
Prints a whaely good message to the console.
}
\author{
\strong{Maintainer}: Mark Hanly \email{m.hanly@unsw.edu.au} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
