% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owhaley.R
\name{say}
\alias{say}
\title{owhaley}
\usage{
say(what)
}
\arguments{
\item{what}{text string; Whatever phrase you want whale to echo}
}
\value{
whale with phrase
}
\description{
summon a whale for customizable pick me up
}
\examples{
say(what = "you're whale-come!")
}
