% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/say.R
\name{say}
\alias{say}
\title{Speaking whale}
\usage{
say(what)
}
\arguments{
\item{what}{What the whale will say. If empty defaults to generic whale comment.}
}
\value{
Whale message to the console
}
\description{
Speaking whale
}
\examples{
say('Well head')
}
