% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/say.R
\name{say}
\alias{say}
\title{Summon a whale for a customisable pick-me up}
\usage{
say(what)
}
\arguments{
\item{what}{whatever phrase you want the whale to echo}
}
\value{
whale shaped message
}
\description{
Summon a whale for a customisable pick-me up
}
\examples{
say("You're whale-come")

}
